Refactor magic number into class method
Use a class method instead of a constant, because if we want this magic
number to be dynamic, we'll need to be able to supply it an argument.
Defining as a class method now means we can change the implementation
down the line with no cost of complexity right now.