Spec out a CreatesProject Workflow
This will be used to handle the business logic of creating a project
and tasks from the params received from the form. We like using
Workflows, cause they're easier to test. It's nice to test unit tests
that wire the end-to-end together, then you can just make sure
everything is working with an integration test at the end of the day.